Product of Array Except Self(Day 16)
I initially started solving this problem by taking the product of all the elements of the array then dividing the product by the i-th pos array element

Obviously problem of 0 came up and it failed

Then i pivoted to basically adding a bunch of if-else statements to cover as many edge cases as i can. Turns out the code gets crazy messsy

Finally i solved the problem by dividing the array into two parts pivoted around the index and multiply the left and right side of that

The main point to remember here is that
if you find yourself storing a lot of info about the process and its individual stages
Such as checking if a zero has occured or not
Or finding yourself in a situation where you think having a history of all occurances of zero would be helpful it is a indication that you need to use array.